fix : delete votescreen file and join to App.js
